Tools/mcp/project_context/requirements.txt
fastmcp
pyyaml
pydantic
